Fixed IMGUI Unusable: Mouse Doesn't Work in Game Launcher https://github.com/o3de/o3de/issues/12932
The problem was introduced at https://github.com/o3de/o3de/pull/12609 which moved ImGuiManager rendering to a job thread, as an optimization. I simply reverted that part of the changes, as there isn't time for a thorough investigation and proper fix, and I'm concerned there might be more threading issues yet to be uncovered. If we really want ImGuiManager to render on a separate thread, someone else will have to do a deep dive.

The original change is reported to have saved 5ms, but only when the ImGui CPU profiler tool was open, and on a specific level. When I tested using the default scene, the improvement was 1ms.

Signed-off-by: santorac <55155825+santorac@users.noreply.github.com>